test: EntityPersister 테스트수정
